Module honeybee::modules::run_watch

Expands the given path to a pidfile and checks if the process ID found inside is of a running process. You can use this to check if a specific application, such as a VPN client, is running.

Configuration

Key	Description	Values	Default

title	Will replace the placeholder $title	A string	N/A

pidfile	A path to a pidfile or a pattern that matches to pathnames (see glob(3))	A string	N/A

format	A custom string to format the status of the process in the bar	A string with placeholders	"$title: $status"

format_down	The same as above for when the process is not running	A string with placeholders	"$title: $status"

Placeholder	Value

$title	A descriptive title for your path (e.g. “VPN”)

$status	A string with the values yes and no, indicating whether the path exists or not

Example

[[module]]
module = "run_watch"
title = "Spotify"
pidfile = "/path/to/some/pidfile"

[[module]]
module = "run_watch"
title = "VPN"
pidfile = "/path/to/vpn*.pid"
format_down = "$title is down!!!"

Structs

Config