honeybee::modules

Module time

Outputs the current time in the local timezone, if timezone is not set, or in given timezone, if timezone is set.

Configuration

Key	Description	Values	Default

title	Identifies this particular time instance (you can have multiple)	A string	N/A

format	Format string used to define what the output will look like	See below	"%Y-%m-%d %H:%M %Z"

timezone	The current time will be output in the given timezone	See below	"Local"

Format

The value of format should be a string with valid identifiers. You can also see strftime(3) for details.

TimeZone

The value of timezone can be Local (meaning that the local time will be used), UTC or any valid TZ identifier.

Example

[[module]]
module = "time"
title = "Amsterdam"
format = "%Y-%m-%d %H:%M %Z"
timezone = "Europe/Amsterdam"

Structs

Config