honeybee::modules

Module wireless

Gets the link signal, frequency and SSID of the given wireless network interface.

Configuration

Key	Description	Values	Default

interface	Wireless network interface to monitor (as specified in /sys/class/net/)	A string	"wlan0"

format_up	String to customize the output of the module when the interface is up. See placeholders below	`“$name: [$signal at $ssid, $bitrate [D: $speed_down	U: $speed_up]] $ipv4“`

format_down	String to customize the output of the module when the interface is down.	"W: down"

Placeholder	Value

$name	Name of the Wireless network interface.

$ipv4	IPv4 address of the interface

$ipv6	IPv6 address of the interface

$ssid	Network SSID

$signal	WiFi signal strength

$bitrate	WiFi connection bitrate

$frequency	Interface frequency of the selected channel (Ghz)

$speed_up	Upload speed

$speed_down	Download speed

Example

[[module]]
module = "wireless"
interface = "wlan0"

Structs

Config