honeybee

Module output

The output_format configuration defines the format strings honeybee will use in its outputs.

Supported output formats are i3bar, dzen2, xmobar, lemonbar and term.

Modules

color

Everything needed to manipulate colors.

dzen2

Dzen is a general purpose messaging, notification and menuing program for X11.

i3bar

i3bar output format

lemonbar

lemonbar is a lightweight bar based entirely on XCB. It has full UTF-8 support and is EWMH compliant.

term

Use ANSI Escape sequences to produce a terminal-output as close as possible to the graphical outputs. This makes debugging your config file a little bit easier because the terminal-output becomes much more readable.

xmobar

xmobar is a minimalistic, mostly text based, status bar. It was originally designed and implemented by Andrea Rossato to work with xmonad, but it’s actually usable with any window-manager.

Traits

OutputFormatter

Functions

output_formatter