honeybee

Module protocol

i3bar protocol. For full reference, check the main documentation.

The main structures of this module are Header, Block and ClickEvent. This is only relevant if you choose [i3bar] as the output_format.

Structs

Block

Represents the contents of the status line for a particular module.

ClickEvent

Header

The first message in the protocol is a header block.

Enums

Alignment

BlockWidth

Markup

Functions

init

status_line

stop