Crate honeybee

honeybee is a port of i3status, written in Rust. It means that it’s compatible with the i3bar protocol, but also generates status line for dzen2,xmobar and lemonbar.

Modules

bar

This is the main container of all the running modules.

bridge

Establishes the communication between individual modules and the Bar. Each module has its own bridge.

config

errors

formatting

Everything related to format strings, placeholders and value types.

modules

output

The output_format configuration defines the format strings honeybee will use in its outputs.

protocol

i3bar protocol. For full reference, check the main documentation.

timer

Used to control the update interval of modules.

types

units

utils

Utilities, mostly to manipulate files.

widget

Structs and methods to manipulate widgets. A Widget is a graphical element that is used to render the output of a Module.

Macros

from_str

map

modules

Structs

CliArgs

A port of i3status, written in Rust.