Module honeybee::modules

Modules

battery

Gets the status (charging, discharging, unknown, full), percentage, remaining time and power consumption (in Watts) of the given battery and when it’s estimated to be empty.

disk

Gets used, free, available and total amount of bytes on the given mounted filesystem.

file_contents

Outputs the contents of the specified file. You can use this to check contents of files on your system, for example /proc/uptime.

load_avg

Gets the system load (number of processes waiting for CPU time in the last 1, 5 and 15 minutes).

memory

Memory usage on a Linux system from /proc/meminfo.

path_exists

Check if the given path exists in the filesystem. You can use this to check if something is active, like for example a VPN connection managed by NetworkManager.

run_watch

Expands the given path to a pidfile and checks if the process ID found inside is of a running process. You can use this to check if a specific application, such as a VPN client, is running.

time

Outputs the current time in the local timezone, if timezone is not set, or in given timezone, if timezone is set.

wireless

Structs

Module

Enums

ModuleConfig