honeybee::modules

Module battery

Gets the status (charging, discharging, unknown, full), percentage, remaining time and power consumption (in Watts) of the given battery and when it’s estimated to be empty.

Configuration

Key	Description	Values	Default

number	Battery index as reported in /sys	An integer	0

path	Set this property if your battery is represented in a non-standard path in /sys. The first occurrence of %d gets replaced with number, but you can also hardcode the battery index in the path	String representing a path	"/sys/class/power_supply/BAT%d/uevent"

format	A string used to customize the output of this module	See available placeholders below	"$status $percentage $remaining"

format_down	A string to customize the output when there are no metrics	A string of UTF-8 characters	"No battery"

status_chr	Custom string to represent charging state.	String with any UTF-8 symbols	"CHR"

status_bat	same as above, but for discharging.	same as above	"BAT"

status_unk	same as above, but for unknown	same as above	"UNK"

status_full	same as above, but for full	same as above	"FULL"

low_threshold	Causes the widget state to change to WidgetState::Critical	Integer	10

threshold_type	An integer representing either time or percentage. If you define percentage and your battery percentage goes below low_threshold, the widget state will change accordingly.	time, percentage	percentage

Placeholder	Value

$status	Status (charging, discharging, unknown, full) of the battery. Will be replaced by one of the status_* properties defined above.

$percentage	Battery left percentage

$remaining	Remaining time

$emptytime	Time when it’s estimated to be empty

$consumption	Power consumption in Watts

Example

[[module]]
module = "battery"
number = 0
format = "$status $percentage $remaining"
format_down = "No battery"
status_chr = "⚇ CHR"
status_bat = "⚡ BAT"
status_full = "☻ FULL"
low_threshold = 20
threshold_type = "percentage"

Structs

Config

Enums

ChargingStatus

Represents the operating status. Corresponds to POWER_SUPPLY_STATUS_*, as defined in power_supply.h.

ThresholdType