Module honeybee::modules::disk

Gets used, free, available and total amount of bytes on the given mounted filesystem.

Configuration

Key	Description	Values	Default

path	Path to the mounted filesystem	A string representing a path	"/"

format	A string used to customize the output of this module	See available placeholders below	"$free"

format_below_threshold	A string to customize the output when state is set to criticalo	See available placeholders below	"Warning: $percentage_avail"

format_not_mounted	A string to customize the output when the path doesn’t exist or is not a mount point.	A string	""

prefix_type	Prefix used to present byte sizes in human readable format	See options below	"binary"

threshold_type	Type of low_threshold that sets the widget state to critical	See options below	"percentage_avail"

low_threshold	Value that causes the disk text to be displayed using color_bad (widget state critical)	Number	0

Prefix Type	Value

binary	IEC prefixes (KiB, MiB, GiB, TiB) represent multiples of powers of 1024

decimal	SI prefixes (K, M, G, T) represent multiples of powers of 1000

Threshold Type	Value

percentage_free	Percentage of disk space free

percentage_avail	Percentage of disk space available

bytes_free	Number of bytes free

bytes_avail	Number of bytes available

Note that bytes_free and bytes_avail can be prefixed with “k”, “m”, “g” or “t”. That means that if you set prefix_type to binary, low_threshold to 2 and threshold_type to gbytes_avail, then the disk info will be colored bad.

If not specified, threshold_type is assumed to be percentage_avail and low_threshold is 0, which implies no coloring at all.

Placeholder	Value

$free	Free disk space

$percentage_free	As above, but percentage

$avail	Available disk space

$percentage_avail	As above, but percentage

$total	Total disk space

$used	Used disk space

$percentage_used	As above, but percentage

$percentage_used_of_avail	Percentage of available space being used

Example

[[module]]
module = "disk"
format = "/: $avail"
prefix_type = "binary"
threshold_type = "gbytes_avail"
low_threshold = 10

Structs

Config